ValueFlow: extracted `valueFlowNumber()` and related common code into separate files
