+++ lib/lib.pri
@@ -78,7 +78,10 @@ HEADERS += $${PWD}/addoninfo.h \
            $${PWD}/valueptr.h \
            $${PWD}/version.h \
            $${PWD}/vf_analyze.h \
+           $${PWD}/vf_common.h \
            $${PWD}/vf_enumvalue.h \
+           $${PWD}/vf_number.h \
+           $${PWD}/vf_settokenvalue.h \
            $${PWD}/vfvalue.h \
            $${PWD}/xml.h
 
@@ -144,5 +147,8 @@ SOURCES += $${PWD}/valueflow.cpp \
            $${PWD}/token.cpp \
            $${PWD}/tokenlist.cpp \
            $${PWD}/utils.cpp \
+           $${PWD}/vf_common.cpp \
            $${PWD}/vf_enumvalue.cpp \
+           $${PWD}/vf_number.cpp \
+           $${PWD}/vf_settokenvalue.cpp \
            $${PWD}/vfvalue.cpp
