+++ lib/vf_number.h
@@ -16,10 +16,15 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef vfAnalyzeH
-#define vfAnalyzeH
+#ifndef vfNumberH
+#define vfNumberH
 
-#include "vf_enumvalue.h" // IWYU pragma: export
-#include "vf_number.h" // IWYU pragma: export
+class TokenList;
+class Settings;
 
-#endif // vfAnalyzeH
+namespace ValueFlow
+{
+    void analyzeNumber(TokenList &tokenlist, const Settings& settings);
+}
+
+#endif // vfNumberH
